vmalert: improve template performance
